requirements.txt
Django==4.0.6
celery==5.2.7 
psycopg2>=2.8
redis==4.3.4
djangorestframework==3.13.1 